Refactor: usePlanStateMutation.ts생성
